fix(groups sidebar): fix failure to rerender on changes to groups
The component wasn't rerendering on a change inside the mobx store, the props now take the goups
through a wrapping component.
